fix : 연동과정에서 발생한 오류들 해결
