+++ src/main/java/com/example/carrotdiary/global/constants/Role.java
@@ -1,6 +1,22 @@
 package com.example.carrotdiary.global.constants;
 
 public enum Role {
-    ADMIN, USER
+    ADMIN(Authority.ADMIN),
+    USER(Authority.USER);
 
+    private final String authority;
+
+    Role(String authority) {
+        this.authority = authority;
+    }
+
+    public String getAuthority() {
+        return this.authority;
+    }
+
+    public static class Authority {
+        public static final String ADMIN = "ROLE_ADMIN";
+        public static final String USER = "ROLE_USER";
+    }
 }
+
