+++ src/main/java/com/example/carrotdiary/global/security/UserDetailsImpl.java
@@ -1,8 +1,7 @@
-package com.example.carrrotdiary.global.security;
+package com.example.carrotdiary.global.security;
 
-import com.example.carrrotdiary.global.constants.Role;
-import com.example.carrrotdiary.member.dto.MemberDetailDto;
-import com.example.carrrotdiary.member.entity.MemberEntity;
+import com.example.carrotdiary.member.dto.MemberDetailDto;
+import com.example.carrotdiary.member.entity.MemberEntity;
 import org.springframework.security.core.userdetails.UserDetails;
 
 
@@ -10,7 +9,6 @@ import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
