--- src/main/java/com/example/carrrotdiary/global/constants/Role.java
@@ -1,22 +1,0 @@
-package com.example.carrrotdiary.global.constants;
-
-public enum Role {
-    ADMIN(Authority.ADMIN),
-    USER(Authority.USER);
-
-    private final String authority;
-
-    Role(String authority) {
-        this.authority = authority;
-    }
-
-    public String getAuthority() {
-        return this.authority;
-    }
-
-    public static class Authority {
-        public static final String ADMIN = "ROLE_ADMIN";
-        public static final String USER = "ROLE_USER";
-    }
-}
-
